pyproject.toml
[tool.poetry]
name = "gpt-serviece-for-engineer"
version = "0.1.0"
description = ""
authors = ["JPark <qkrwogus88@naver.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.17"
langchain-core = "^0.1.50"
langchain-community = "^0.0.36"
openpyxl = "^3.1.2"
pandas = "^2.2.2"
tiktoken = "^0.6.0"
openai = "^1.25.1"
langchain-openai = "^0.1.6"
pymupdf = "^1.24.2"
python-dotenv = "^1.0.1"
umap = "^0.1.1"
scikit-learn = "^1.4.2"
duckduckgo-search = "^5.3.0"
numexpr = "^2.10.0"
markdown = "^3.6"
faiss-cpu = "^1.8.0"
streamlit = "^1.34.0"
streamlit-book = "^0.7.5"
st-pages = "^0.4.5"
pillow = "^10.3.0"
reportlab = "^4.2.0"
pymupdf4llm = "^0.0.3"
markdown2 = "^2.4.13"
python-docx = "^1.1.2"
streamlit-feedback = "^0.1.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
